#organized js codes, remove extra codes, and add send message backend methods;
