+++ src/main/java/com/issa/kafka/master/controllers/MessageController.java
@@ -2,15 +2,18 @@ package com.issa.kafka.master.controllers;
 
 import com.issa.kafka.master.dto.filters.GetAllMessagesFilter;
 import com.issa.kafka.master.dto.filters.TopicServerNameFilter;
+import com.issa.kafka.master.dto.forms.SendMessageForm;
 import com.issa.kafka.master.enums.ServiceResultStatus;
 import com.issa.kafka.master.services.MessageService;
 import com.issa.kafka.master.utility.ResponseResult;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+import java.util.Map;
 
 @Controller
 @RequiredArgsConstructor
@@ -54,6 +57,39 @@ public class MessageController {
 
             return ResponseEntity.ok(getAllMessageResponseResult);
 
+        } catch (Exception e) {
+            return ResponseEntity.ok(
+                    new ResponseResult(
+                            ServiceResultStatus.ERROR,
+                            false));
+        }
+    }
+
+
+    @PostMapping("/sendsingle")
+    public ResponseEntity<ResponseResult> sendMessage(@RequestBody SendMessageForm sendMessageForm) {
+        try {
+            ResponseResult sendMessageResponseResult =
+                    messageService.sendMessage(sendMessageForm.getTopicName(), sendMessageForm.getServerName(),
+                            sendMessageForm.getPartition(), sendMessageForm.getMessage().get(0));
+
+            return ResponseEntity.ok(sendMessageResponseResult);
+        } catch (Exception e) {
+            return ResponseEntity.ok(
+                    new ResponseResult(
+                            ServiceResultStatus.ERROR,
+                            false));
+        }
+    }
+
+    @PostMapping("/sendmultiple")
+    public ResponseEntity<ResponseResult> sendMessages(@RequestBody SendMessageForm sendMessageForm) {
+        try {
+            ResponseResult sendMessageResponseResult =
+                    messageService.sendMessages(sendMessageForm.getTopicName(), sendMessageForm.getServerName(),
+                            sendMessageForm.getPartition(), sendMessageForm.getMessage());
+
+            return ResponseEntity.ok(sendMessageResponseResult);
         } catch (Exception e) {
             return ResponseEntity.ok(
                     new ResponseResult(
