+++ src/main/resources/server-status-message.properties
@@ -3,11 +3,12 @@ FIELDS_REQUIRED=Fields are required.
 SERVER_NAME_DUPLICATE=Server name already exists. Please choose a different name.
 SERVER_ID_DUPLICATE=Server ID already exists. Please check the IP and port.
 CHECK_CONNECTION_PARAMS=check if kafka server available.
+NO_CONNECTION_FOUND=not found current connection.
 
 
 DONE=Operation done successfully.
 ERROR=An error occurred during the operation.
-
+PENDING=operation pending.
 
 NO_CLIENT_FOUND=No Kafka client found. Try to reconnect.
 
@@ -21,6 +22,8 @@ FAIL_TO_GET_MESSAGES=failed to get messages.
 FAIL_TO_CREATE_TOPIC=failed to create topic.
 TOPIC_NAME_DUPLICATE=topic name is already exist.
 FAIL_TO_DELETE_TOPIC=failed to delete topic.
+TOPIC_NOT_FOUND=selected topic not found.
+FAIL_TO_SEND_MESSAGE=error in publishing message on topic.
 
 #filter status messages
 FROM_DATE_INVALID=date 'FROM' is invalid.
