+++ src/main/resources/templates/fragments/topic-modal.html
@@ -5,7 +5,7 @@
     <div class="tab-content" id="topic-tabs" style="display:none;">
         <ul class="nav nav-tabs">
             <li class="nav-item">
-                <a class="nav-link active" id="properties-tab" data-bs-toggle="tab" href="#properties" role="tab">Properties</a>
+                <a class="nav-link active" id="properties-tab" data-bs-toggle="tab" href="#properties" role="tab" onclick="refreshMessagesCount()">Properties</a>
             </li>
             <li class="nav-item">
                 <a class="nav-link" id="data-tab" data-bs-toggle="tab" href="#data" role="tab">Data</a>
