+++ src/main/resources/templates/main-page.html
@@ -56,6 +56,8 @@
 <script th:src="@{/js/common.js}"></script>
 <script th:src="@{/js/consumer-groups.js}"></script>
 <script th:src="@{/js/topics.js}"></script>
+<script th:src="@{/js/partitions.js}"></script>
+<script th:src="@{/js/configs.js}"></script>
 <script th:src="@{/js/connect.js}"></script>
 <script th:src="@{/js/messages.js}"></script>
